example/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
googleapis-common-protos==1.63.1rc0
grpcio==1.64.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
opentelemetry-api==1.25.0
opentelemetry-distro==0.46b0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-arangodb==0.1.0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-asyncio==0.46b0
opentelemetry-instrumentation-aws-lambda==0.46b0
opentelemetry-instrumentation-dbapi==0.46b0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-jinja2==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-sqlite3==0.46b0
opentelemetry-instrumentation-threading==0.46b0
opentelemetry-instrumentation-urllib==0.46b0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-test-utils==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.0
protobuf==4.25.3
PyJWT==2.8.0
python-arango==8.0.0
requests==2.32.3
requests-toolbelt==1.0.0
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.1
